+++ spec/selector-solver-spec.js
@@ -59,8 +59,13 @@ describe("Tag solver", function () {
 		expect(solver.solveSelector("div#mama", true, true, true, ['div']).string).toBe('<div id="mama">\n\n</div>');
 	});
 
-	it("shouldn't expand tag if class is not specified", function () {
+	it("shouldn't expand tag if class or id is not specified", function () {
+		expect(solver.solveSelector("div#")).toBe(null);
 		expect(solver.solveSelector("div.")).toBe(null);
 	});
 
+	it("should solve to div if only class or id specified", function () {
+		expect(solver.solveSelector(".mama").string).toBe('<div class="mama"></div>');
+		expect(solver.solveSelector("#tata").string).toBe('<div id="tata"></div>');
+	});
 })